LPS-60768 Extract upgrade process to main registry.
With the new pattern, UpgradeJournalArticleType runs always before UpgradeJournalArticles, so there's no need to nest the second inside the first to force this sequence.
